+++ iModelBridgeCore/Imagepp/all/gra/hrf/src/HRFBilLineEditor.cpp
@@ -170,7 +170,7 @@ HSTATUS HRFBilLineEditor::Read24BitRgbBlock(uint64_t pi_PosBlockX,
 
     if (GetResolutionDescriptor()->GetScanlineOrientation().IsLeft())
         {
-        for (PixelIndex = 0; PixelIndex < m_LineWidth; PixelIndex++)
+        for (PixelIndex = 0; PixelIndex < (int32_t) m_LineWidth; PixelIndex++)
             {
             po_pData[BytePos++]= m_pRedLineBuffer  [PixelIndex];
             po_pData[BytePos++]= m_pGreenLineBuffer[PixelIndex];
