+++ iModelBridgeCore/Imagepp/all/gra/hrf/src/HRFDoqEditor.cpp
@@ -175,7 +175,7 @@ HSTATUS HRFDoqEditor::Read24BitRgbBlock(uint64_t pi_PosBlockX,
 
     if (GetResolutionDescriptor()->GetScanlineOrientation().IsLeft())
         {
-        for (PixelIndex = 0; PixelIndex < m_LineWidth; PixelIndex++)
+        for (PixelIndex = 0; PixelIndex < (int32_t) m_LineWidth; PixelIndex++)
             {
 
             po_pData[BytePos++]= m_pRedLineBuffer  [PixelIndex];
