+++ iModelBridgeCore/Imagepp/all/gra/hrf/src/HRFTWFPageFile.cpp
@@ -777,7 +777,7 @@ void HRFTWFPageFile::CleanUpString(string* pio_pString) const
     if(!pio_pString->empty())
         {
         // Remove the SPACE/TAB/ENTER at the end of the string.
-        Pos = pio_pString->size() - 1;
+        Pos = (int32_t) pio_pString->size() - 1;
         while(Pos >= 0 && !IsValidChar((*pio_pString)[Pos]))
             Pos--;
