+++ iModelBridgeCore/Imagepp/all/gra/hrp/src/HRPDitherFilter.cpp
@@ -386,10 +386,9 @@ void HRPDitherFilter::FillQuantizedPalette()
     const HRPPixelPalette& rPalette = m_pPixelType->GetPalette();
 
     // fill the octree with the destination palette entries
-//&&ep - check if was changed recently. Could not work.    for(Byte Index = 0; Index < 256; Index++)
     for(uint16_t Index = 0; Index < 256; Index++)
         {
         m_QuantizedPalette.AddCompositeValue(rPalette.GetCompositeValue(Index),
-                                             Index);
+                                             (Byte)Index);
         }
     }
